apache-solr::setup removed trailing slash on solr_base
only remove schema.xml if file exists

+++ apache-solr/recipes/setup.rb
@@ -1,7 +1,7 @@
 # install the example application and prepare it
 
 solr_app  = node[:apache_solr][:app]
-solr_base = "#{node[:apache_solr][:base_dir]}/apache-solr/"
+solr_base = "#{node[:apache_solr][:base_dir]}/apache-solr"
 
 if !File.directory?("#{solr_base}/#{solr_app}")
 
@@ -16,6 +16,9 @@ if !File.directory?("#{solr_base}/#{solr_app}")
   # this should keep solr from starting and error accordingly ;)
   execute "delete standard schema.xml" do
     cwd     "#{solr_base}/#{solr_app}/solr/conf"
+    only_if do
+      File.exists?("#{solr_base}/#{solr_app}/solr/conf/schema.xml")
+    end
     command "rm schema.xml"
   end
